requirements.txt
requests==2.31.0
praw==7.7.1
kafka-python==2.0.2
schedule==1.2.1
transformers==4.40.0
torch==2.3.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
loguru==0.7.2
streamlit 
plotly
pytest